Update a choose_order test, and use Float64 for loop length product calculations due to larger range.

+++ src/lower_load.jl
@@ -86,7 +86,7 @@ function prefetchisagoodidea(ls::LoopSet, op::Operation, td::UnrollArgs)
             innermostloopindv = findall(map(isequal(innermostloopsym), getindices(op)))
             isone(length(innermostloopindv)) || return 0
             innermostloopind = first(innermostloopindv)
-            if prod(s -> length(getloop(ls, s)), @view(indices[1:innermostloopind-1])) ≥ 120 && length(getloop(ls, innermostloopsym)) ≥ 120
+            if prod(s -> Float64(length(getloop(ls, s))), @view(indices[1:innermostloopind-1])) ≥ 120.0 && length(getloop(ls, innermostloopsym)) ≥ 120
                 if op.ref.ref.offsets[innermostloopind] < 120
                     for opp ∈ operations(ls)
                         iscompute(opp) && (innermostloopsym ∈ loopdependencies(opp)) && load_constrained(opp, u₁loopsym, u₂loopsym, true) && return 0
